+++ b2sdk/sync/folder.py
@@ -317,8 +317,8 @@ class B2Folder(AbstractFolder):
         """
         current_name = None
         current_versions = []
-        for (file_version_info, folder_name) in self.bucket.ls(
-            self.folder_name, show_versions=True, recursive=True, fetch_count=1000
+        for file_version_info, _ in self.bucket.ls(
+            self.folder_name, show_versions=True, recursive=True
         ):
             assert file_version_info.file_name.startswith(self.prefix)
             if file_version_info.action == 'start':