+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add support for Python 3.9
 * Support for bucket to bucket sync
 
+### Changed
+* Change default fetch count for `ls` to 10000
+
 ### Removed
 * Drop Python 2 and Python 3.4 support :tada:
 * Remove `--prefix` from `ls` (it didn't really work, use `folderName` argument)